module7-task2 only one point to set city and places
