+++ src/store/all-hotels/reducers.js
@@ -1,7 +1,6 @@
 import {ActionCreator, ActionType} from "./actions";
 import {AllHotels} from "../../models/all-hotels";
 import {Operation as PlacesInCityOperation} from "../places-in-city/reducers";
-import {ActionCreator as FiltersActionCreator} from "../filters/actions";
 
 const initialState = {
   hotels: []
@@ -15,7 +14,6 @@ const Operation = {
         dispatch(ActionCreator.loadAllHotels(hotels));
 
         const cityName = hotels[0].city.name;
-        dispatch(FiltersActionCreator.setCurrentCityName(cityName));
         dispatch(PlacesInCityOperation.loadPlacesInCity(cityName));
       });
   }
