+++ src/store/places-in-city/actions.js
@@ -1,4 +1,5 @@
 const ActionType = {
+  // SET_CURRENT_CITY: `SET_CURRENT_CITY`,
   SET_PLACES_IN_CITIES: `LOAD_PLACES_IN_CITIES`
 };
